Refactor part2
