+++ src/main/java/com/infoshareacademy/drinkers/App.java
@@ -1,5 +1,11 @@
 package com.infoshareacademy.drinkers;
 
+import com.infoshareacademy.drinkers.domain.drink.Drink;
+import com.infoshareacademy.drinkers.service.filtering.FilterElements;
+import com.infoshareacademy.drinkers.service.filtering.FilterList;
+import com.infoshareacademy.drinkers.service.sorting.SortDrinks;
+import com.infoshareacademy.drinkers.service.sorting.SortItems;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
