+++ src/main/java/com/infoshareacademy/drinkers/DrinkManager.java
@@ -1,5 +1,7 @@
 package com.infoshareacademy.drinkers;
 
+import com.infoshareacademy.drinkers.domain.drink.Drink;
+
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
