+++ src/main/java/com/infoshareacademy/drinkers/service/console/ConsoleInput.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.drinkers.service;
+package com.infoshareacademy.drinkers.service.console;
 
 import java.util.InputMismatchException;
 import java.util.Scanner;
