+++ src/main/java/com/infoshareacademy/drinkers/service/filtering/FilterElements.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.drinkers;
+package com.infoshareacademy.drinkers.service.filtering;
 
 public enum FilterElements {
     ALCOHOL("Alcoholic"),
