+++ src/main/java/com/infoshareacademy/drinkers/service/sorting/SortItems.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.drinkers;
+package com.infoshareacademy.drinkers.service.sorting;
 
 public enum SortItems {
     ID("ID"),
